Objectives achieved: Share poll link after creating new poll. Home tab will show polls created by user only. From home screen polls can be closed or shared. Once the poll is closed no one can vote. And poll will be moved to Result tab. Reduced duration of animation of newPoll fab.
ToDo:
Add a ResultActivity which will launch on clicking "Show Result" in Results Tab.
Calculate winner and show it on this activity.
Add a toggle button to switch between "myPolls" and to the polls user did vote for.
Add a sharing option in ResultActivity for sharing the result (propbably in ScreenShot form).
Improve UI.
Testing.